ToDo Program
Added a todo program as well as built a namespace in a header file that resembles functions present in microsoft small basic. Also added to a currently delyed project atari adventure. This delay is due to me working on the small basic library header as well as my todo list program.