Decals PGEX
Made an animation managment header based on the pge extention call olcPGEX_Graphics2D.h
The problem with this header is the fact that it didnt support decals so i made my own implementation which took me over 5 hours for a 6 line fix for  a bug that prevented anything from being drawn. Time to go and cry myslef to sleep.