Refactorisation
Refactored the data structure for the classes so that i could wrap up  100's of lines of code. Made it so you could pass i n any object and it would read its discription and check if it was clicked instead of manually checking for every object type for each gamestate and then checking if you were clicking it. This saved an estimated 200 lines give or take because other functions benifited from the more simplistic refacorisation.