Many New Stuff
I forgot to add crap to the github so here it is in a .7z file that contains 7 commits that i didnt commit. Anyway i did alot and the system is working great. The best develpments is word wraping and windows CLI functions that ive added into the program to make everything look perty.